Including the files from MessagePack-CSharp
instead of referencing the NuGet package. The reason for this is that in future releases the class MessagePackBinary isn't present, it is replaced with MessagePackReader, which is a ref struct, and therefore can't be stored as a field of a class. Therefore we won't be able to implement these JsonReader/JsonWriter classes in a future release. This sucks but perhaps it's the best solution we have.